+++ src/vs/platform/agentHost/node/copilot/copilotSessionLauncher.ts
@@ -131,10 +131,12 @@ interface ICopilotSessionLaunchBase {
 	readonly githubToken: string | undefined;
 
 	/**
-	 * Whether this is a workspace-less quick chat. Threaded into the prompt
-	 * context so the resolved system message gets the scratch/repoless variant.
+	 * Whether this is a workspace-less session (a quick chat). Threaded into the
+	 * prompt context so the resolved system message gets the scratch/repoless
+	 * variant. Named to match the `workspaceless` marker used throughout the AH
+	 * layer (session `_meta`, stored metadata) that this value flows from.
 	 */
-	readonly isQuickChat?: boolean;
+	readonly workspaceless?: boolean;
 }
 
 export interface ICopilotCreateSessionLaunchPlan extends ICopilotSessionLaunchBase {
@@ -491,7 +493,7 @@ export class CopilotSessionLauncher implements ICopilotSessionLauncher {
 		const promptContext: IAgentHostPromptContext = {
 			getSetting: key => this._configurationService.getRootValue(agentHostCustomizationConfigSchema, key),
 			hasClientTool: name => clientToolNames.has(name),
-			isQuickChat: plan.isQuickChat === true,
+			workspaceless: plan.workspaceless === true,
 		};
 		// Resolved once per (re)launch — the SDK has no mid-session system-message
 		// update, so this reflects the model/tools/settings at launch time. Log a
