+++ src/vs/platform/agentHost/test/node/agentHostPromptRegistry.test.ts
@@ -18,12 +18,12 @@ import '../../node/copilot/prompts/allPrompts.js';
  * Builds a prompt context backed by an in-memory bag of customization settings
  * and an optional set of available tool names.
  */
-function context(settings: SchemaValues<typeof agentHostCustomizationConfigSchema.definition> = {}, tools: readonly string[] = [], isQuickChat = false): IAgentHostPromptContext {
+function context(settings: SchemaValues<typeof agentHostCustomizationConfigSchema.definition> = {}, tools: readonly string[] = [], workspaceless = false): IAgentHostPromptContext {
 	const toolNames = new Set(tools);
 	return {
 		getSetting: key => settings[key],
 		hasClientTool: name => toolNames.has(name),
-		isQuickChat,
+		workspaceless,
 	};
 }
